Moved signature generations methods to new TypeDefinitionGenerator
